requirements-dev.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
Pillow==8.3.2
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
